requirements.txt
streamlit
PyMuPDF
resume-parser
google-generativeai
python-dotenv